+++ app/src/main/java/com/example/langbridge/login/ui/LoginViewModel.kt
@@ -5,13 +5,34 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.langbridge.login.data.models.LoginResponse
+import com.example.langbridge.login.data.models.SignInResult
+import com.example.langbridge.login.data.models.SignInState
 import com.example.langbridge.login.data.repository.LoginRepository
 import com.example.langbridge.login.data.repository.LoginRepositoryImpl
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 
 class LoginViewModel : ViewModel() {
+
+    private val _state = MutableStateFlow(SignInState())
+    val state = _state.asStateFlow()
+
+    fun onSignInResult(result: SignInResult) {
+        _state.update { it.copy(
+            isSignInSuccessful = result.data != null,
+            signInError = result.errorMessage
+        ) }
+    }
+
+    fun resetState() {
+        _state.update { SignInState() }
+    }
+
+
     var name = mutableStateOf("")
     var error = mutableStateOf(false)
     var message = mutableStateOf("")
