--- app/src/main/java/com/example/langbridge/loginWithGoogle/GoogleSignInScreen.kt
@@ -1,42 +1,0 @@
-package com.example.langbridge.loginWithGoogle
-
-import android.widget.Toast
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material.Button
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.unit.dp
-
-@Composable
-fun GoogleSignInScreen(
-    state: SignInState,
-    onGoogleSignInClick: () -> Unit
-){
-    val context = LocalContext.current
-    LaunchedEffect(key1 = state.signInError) {
-        state.signInError?.let { error ->
-            Toast.makeText(
-                context,
-                error,
-                Toast.LENGTH_LONG
-            ).show()
-        }
-    }
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(16.dp),
-        contentAlignment = Alignment.Center
-    ){
-        Button(onClick = onGoogleSignInClick) {
-            Text(text = "Sign In With Google")
-        }
-    }
-}
