--- app/src/main/java/com/example/langbridge/loginWithGoogle/ProfileScreen.kt
@@ -1,48 +1,0 @@
-package com.example.langbridge.loginWithGoogle
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.height
-import androidx.compose.material3.Button
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-
-@Composable
-fun ProfileScreen(
-    userData: UserData?,
-    onSignOut: () -> Unit
-){
-    Column (
-        modifier = Modifier.fillMaxSize(),
-        verticalArrangement = Arrangement.Center
-    ) {
-        if(userData?.userEmail != null){
-            Text(text = userData.userEmail,
-                textAlign = TextAlign.Center,
-                fontSize = 35.sp,
-                fontWeight = FontWeight.Bold
-            )
-            Spacer(modifier = Modifier.height(16.dp))
-        }
-        if(userData?.username != null){
-            Text(text = userData.username,
-                textAlign = TextAlign.Center,
-                fontSize = 35.sp,
-                fontWeight = FontWeight.Bold
-            )
-            Spacer(modifier = Modifier.height(16.dp))
-        }
-        Button(onClick = onSignOut) {
-            Text(text = "Sign Out")
-        }
-    }
-}
-
-
